requirements.txt
flask
python-socketio
eventlet
gunicorn